refactor: TicketingInfo -> Ticket으로 변경, payment 삭제
