+++ app/src/main/java/woowacourse/movie/activity/ticketresult/TicketResultActivity.kt
@@ -6,7 +6,8 @@ import android.widget.TextView
 import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import woowacourse.movie.R
-import woowacourse.movie.domain.Price
+import woowacourse.movie.domain.ticket.Price
+import woowacourse.movie.domain.ticket.Ticket
 import woowacourse.movie.util.getSerializableExtraCompat
 import java.text.DecimalFormat
 import java.time.LocalDate
@@ -20,7 +21,7 @@ class TicketResultActivity : AppCompatActivity() {
 
         supportActionBar?.setDisplayHomeAsUpEnabled(true)
 
-        val info: woowacourse.movie.domain.TicketingInfo? = intent.getSerializableExtraCompat(INFO_KEY)
+        val info: Ticket? = intent.getSerializableExtraCompat(INFO_KEY)
 
         if (info == null) {
             Toast.makeText(this, DATA_LOADING_ERROR_MESSAGE, Toast.LENGTH_LONG).show()
