+++ app/src/main/res/values/strings.xml
@@ -7,7 +7,7 @@
     <string name="plus">+</string>
     <string name="complete_ticketing">예매 완료</string>
     <string name="cancel_notice">영화 상영 시작 시간 15분 전까지\n취소가 가능합니다.</string>
-    <string name="price_payment">%s원 (%s 결제)</string>
+    <string name="price_payment">%s원 (현장 결제)</string>
     <string name="normal_count">일반 %d명</string>
     <string name="date_time">%s %s</string>
     <string name="date_format">YYYY.M.d</string>
