+++ domain/src/test/java/woowacourse/movie/domain/TicketTest.kt
@@ -6,10 +6,12 @@ import woowacourse.movie.domain.policy.DiscountPolicy
 import woowacourse.movie.domain.policy.MorningPolicy
 import woowacourse.movie.domain.policy.MovieDayPolicy
 import woowacourse.movie.domain.policy.NightPolicy
+import woowacourse.movie.domain.ticket.Price
+import woowacourse.movie.domain.ticket.Ticket
 import java.time.LocalDate
 import java.time.LocalTime
 
-class TicketingInfoTest {
+class TicketTest {
     @Test
     fun `조조에 해당하지만 무비데이면 무비데이 할인이 선적용된다`() {
         val policies = listOf(
@@ -38,6 +40,6 @@ class TicketingInfoTest {
         assertEquals(actual, expected)
     }
 
-    private fun TicketingInfo(policies: List<DiscountPolicy>, playingDate: LocalDate, playingTime: LocalTime, price: Price): TicketingInfo =
-        TicketingInfo.of(policies, "해리포터와 마법사의 돌", playingDate, playingTime, 1, price, "현장")
+    private fun TicketingInfo(policies: List<DiscountPolicy>, playingDate: LocalDate, playingTime: LocalTime, price: Price): Ticket =
+        Ticket.of(policies, "해리포터와 마법사의 돌", playingDate, playingTime, 1, price)
 }
